requirements.txt
django
django-rest-framework
google-cloud-aiplatform
vertexai
